Move Direction to types module
It belongs naturally there as a fundamental entity
